committing before blowing up initialize function
