+++ initializeColors.js
@@ -1,174 +1,28 @@
 const Color = require('colorjs.io').default;
+const { okhsl_to_srgb } = require('./okhsl');
 
-// best known packing of 0 - 15 centerpoints of a sphere in a cube
-// from http://hydra.nat.uni-magdeburg.de/packing/scu/scu.html
-//
-// index = number of points/spheres
-// coordinates are normalized to -0.5 -> 0.5
-const pointsLookup = [
-    [],
-    [[0, 0, 0]],
-    [
-        [-0.18301, -0.18301, -0.18301],
-        [0.18301, 0.18301, 0.18301],
-    ],
-    [
-        [-0.20711, -0.20711, -0.20711],
-        [0.20711, 0.20711, -0.20711],
-        [0.20711, -0.20711, 0.20711],
-    ],
-    [
-        [-0.20711, -0.20711, -0.20711],
-        [0.20711, 0.20711, -0.20711],
-        [0.20711, -0.20711, 0.20711],
-        [-0.20711, 0.20711, 0.20711],
-    ],
-    [
-        [-0.23607, -0.23607, -0.23607],
-        [0, 0.23607, -0.23607],
-        [0.23607, -0.23607, 0],
-        [-0.23607, 0, 0.23607],
-        [0.23607, 0.23607, 0.23607],
-    ],
-    [
-        [0.12132, -0.24264, -0.24264],
-        [-0.24264, 0.12132, -0.24264],
-        [0.24264, 0.24264, -0.12132],
-        [-0.24264, -0.24264, 0.12132],
-        [0.24264, -0.12132, 0.24264],
-        [-0.12132, 0.24264, 0.24264],
-    ],
-    [
-        [0.10388, -0.24986, -0.24986],
-        [-0.24986, 0.10388, -0.24986],
-        [0.23336, 0.23336, -0.24986],
-        [-0.24986, -0.24986, 0.10388],
-        [0.23336, -0.24986, 0.23336],
-        [-0.24986, 0.23336, 0.23336],
-        [0.24986, 0.24986, 0.24986],
-    ],
-    [
-        [-0.25, -0.25, -0.25],
-        [0.25, -0.25, -0.25],
-        [-0.25, 0.25, -0.25],
-        [0.25, 0.25, -0.25],
-        [-0.25, -0.25, 0.25],
-        [0.25, -0.25, 0.25],
-        [-0.25, 0.25, 0.25],
-        [0.25, 0.25, 0.25],
-    ],
-    [
-        [-0.26795, -0.26795, -0.26795],
-        [0.26795, -0.26795, -0.26795],
-        [-0.26795, 0.26795, -0.26795],
-        [0.26795, 0.26795, -0.26795],
-        [0.0, 0.0, 0.0],
-        [-0.26795, -0.26795, 0.26795],
-        [0.26795, -0.26795, 0.26795],
-        [-0.26795, 0.26795, 0.26795],
-        [0.26795, 0.26795, 0.26795],
-    ],
-    [
-        [-0.28571, -0.28571, -0.28571],
-        [0.14286, -0.28571, -0.28571],
-        [-0.28571, 0.28571, -0.28571],
-        [0.14286, 0.28571, -0.28571],
-        [-0.14286, 0.0, 0.0],
-        [0.28571, 0.0, 0.0],
-        [-0.28571, -0.28571, 0.28571],
-        [0.14286, -0.28571, 0.28571],
-        [-0.28571, 0.28571, 0.28571],
-        [0.14286, 0.28571, 0.28571],
-    ],
-    [
-        [-0.29238, -0.29238, -0.29238],
-        [0.03845, -0.03617, -0.29238],
-        [-0.29238, 0.29238, -0.29238],
-        [0.29238, 0.29238, -0.29238],
-        [0.29238, -0.29238, -0.08669],
-        [-0.25419, 0.0, 0.0],
-        [0.03845, 0.29238, 0.03617],
-        [-0.29238, -0.29238, 0.29238],
-        [0.12287, -0.29238, 0.29238],
-        [0.29238, 0.08669, 0.29238],
-        [-0.29238, 0.29238, 0.29238],
-    ],
-    [
-        [0.29274, -0.29289, -0.29289],
-        [-0.29289, -0.20711, -0.29289],
-        [-0.29289, 0.20711, -0.29289],
-        [0.29274, 0.29289, -0.29289],
-        [-0.00553, -0.28721, -0.00553],
-        [-0.00553, 0.28721, -0.00553],
-        [0.29289, 0.0, -0.0],
-        [-0.29289, -0.29289, 0.29274],
-        [-0.29289, 0.29289, 0.29274],
-        [0.29289, -0.29289, 0.29289],
-        [-0.0, 0.0, 0.29289],
-        [0.29289, 0.29289, 0.29289],
-    ],
-    [
-        [-0.29289, -0.29289, -0.29289],
-        [0.29289, -0.29289, -0.29289],
-        [0.0, 0.0, -0.29289],
-        [-0.29289, 0.29289, -0.29289],
-        [0.29289, 0.29289, -0.29289],
-        [0.0, -0.29289, 0.0],
-        [0.29289, 0.0, 0.0],
-        [0.0, 0.29289, 0.0],
-        [-0.29289, -0.29289, 0.29289],
-        [0.29289, -0.29289, 0.29289],
-        [0.0, 0.0, 0.29289],
-        [-0.29289, 0.29289, 0.29289],
-        [0.29289, 0.29289, 0.29289],
-    ],
-    [
-        [-0.29289, -0.29289, -0.29289],
-        [0.29289, -0.29289, -0.29289],
-        [0.0, 0.0, -0.29289],
-        [-0.29289, 0.29289, -0.29289],
-        [0.29289, 0.29289, -0.29289],
-        [0.0, -0.29289, 0.0],
-        [-0.29289, 0.0, 0.0],
-        [0.29289, 0.0, 0.0],
-        [0.0, 0.29289, 0.0],
-        [-0.29289, -0.29289, 0.29289],
-        [0.29289, -0.29289, 0.29289],
-        [0.0, 0.0, 0.29289],
-        [-0.29289, 0.29289, 0.29289],
-        [0.29289, 0.29289, 0.29289],
-    ],
-    [
-        [-0.30769, -0.30769, -0.30769],
-        [0.07692, -0.30769, -0.30769],
-        [0.30769, 0.0, -0.30769],
-        [-0.30769, 0.07692, -0.30769],
-        [0.0, 0.30769, -0.30769],
-        [0.30769, 0.30769, -0.07692],
-        [0.30769, -0.30769, 0.0],
-        [-0.30769, 0.30769, 0.0],
-        [0.00145, 0.00145, 0.00145],
-        [-0.30769, -0.30769, 0.07692],
-        [0.0, -0.30769, 0.30769],
-        [0.30769, -0.07692, 0.30769],
-        [-0.30769, 0.0, 0.30769],
-        [-0.07692, 0.30769, 0.30769],
-        [0.30769, 0.30769, 0.30769],
-    ],
-];
+const randomColorInRange = (hueRange, chromaRange, luminosityRange) => {
+    let h, c, l;
+    do {
+        h = Math.random();
+    } while (!isInRange(h, hueRange));
+    do {
+        c = Math.random();
+    } while (!isInRange(c, chromaRange));
+    do {
+        l = Math.random();
+    } while (!isInRange(l, luminosityRange));
+    return new Color('srgb', okhsl_to_srgb(h, c, l));
+}
 
-const initializeColors = (n) => {
+const initializeColors = (state, config) => {
     const colors = [];
-    const points = pointsLookup[n];
-    for (let i = 0; i < n; i++) {
-        const x = points[i][0] + 0.5;
-        const y = points[i][1] + 0.5;
-        const z = points[i][2] + 0.5;
-        colors.push(new Color('srgb', [x, y, z]));
+    for (let i = 0; i < config.colorCount - state.colors.length; i++) {
+        colors.push(randomColor());
     }
-    return colors;
 };
 
 module.exports = {
+    randomColorInRange,
     initializeColors,
 };
