+++ utils.js
@@ -73,6 +73,20 @@ const cartesianDistance = (a, b) => {
     );
 };
 
+const isInRange = (n, ranges) => {
+    // check if ranges is a single array or an array of ranges
+    if (!Array.isArray(ranges[0])) {
+        ranges = [ranges];
+    }
+    const sortedRanges = ranges.map(range => range.sort((a, b) => a - b));
+    for (let i = 0; i < ranges.length; i++) {
+        if (sortedRanges[i][0] <= n && n <= sortedRanges[i][1]) {
+            return true;
+        }
+    }
+    return false;
+}
+
 
 module.exports = {
     clamp,
@@ -84,4 +98,5 @@ module.exports = {
     cylindricalToCartesian,
     cartesianToCylindrical,
     cartesianDistance,
+    isInRange
 };
